Refactor: Merge Plist and PropertyList into PropertyList
