+++ Sources/Xcodeproj/XcodeProjectModelSerialization.swift
@@ -566,79 +566,3 @@ extension PropertyListSerializable {
     }
 }
 
-/// A very simple representation of a property list.  Note that the `identifier`
-/// enum is not strictly necessary, but useful to semantically distinguish the
-/// strings that represents object identifiers from those that are just data.
-public enum PropertyList {
-    case identifier(String)
-    case string(String)
-    case array([PropertyList])
-    case dictionary([String: PropertyList])
-}
-
-/// Private struct to generate indentation strings.
-fileprivate struct Indentation: CustomStringConvertible {
-    var level: Int = 0
-    mutating func increase() {
-        level += 1
-        precondition(level > 0, "indentation level overflow")
-    }
-    mutating func decrease() {
-        precondition(level > 0, "indentation level underflow")
-        level -= 1
-    }
-    var description: String {
-        return String(repeating: "   ", count: level)
-    }
-}
-
-/// Private function to generate OPENSTEP-style plist representation.
-fileprivate func generatePlistRepresentation(plist: PropertyList, indentation: Indentation) -> String {
-    // Do the appropriate thing for each type of plist node.
-    switch plist {
-
-      case .identifier(let ident):
-        // FIXME: we should assert that the identifier doesn't need quoting
-        return ident
-
-      case .string(let string):
-        return "\"" + Plist.escape(string: string) + "\""
-
-      case .array(let array):
-        var indent = indentation
-        var str = "(\n"
-        indent.increase()
-        for item in array {
-            str += "\(indent)\(generatePlistRepresentation(plist: item, indentation: indent)),\n"
-        }
-        indent.decrease()
-        str += "\(indent))"
-        return str
-
-      case .dictionary(let dict):
-        var indent = indentation
-        let dict = dict.sorted(by: {
-            // Make `isa` sort first (just for readability purposes).
-            switch ($0.key, $1.key) {
-              case ("isa", "isa"): return false
-              case ("isa", _): return true
-              case (_, "isa"): return false
-              default: return $0.key < $1.key
-            }
-        })
-        var str = "{\n"
-        indent.increase()
-        for item in dict {
-            str += "\(indent)\(item.key) = \(generatePlistRepresentation(plist: item.value, indentation: indent));\n"
-        }
-        indent.decrease()
-        str += "\(indent)}"
-        return str
-    }
-}
-
-extension PropertyList: CustomStringConvertible {
-    public var description: String {
-        return generatePlistRepresentation(plist: self, indentation: Indentation())
-    }
-}
